+++ pkg/cookies/cookies.go
@@ -14,7 +14,7 @@ import (
 
 // MakeCookieFromOptions constructs a cookie based on the given *options.CookieOptions,
 // value and creation time
-func MakeCookieFromOptions(req *http.Request, name string, value string, opts *options.Cookie, expiration time.Duration) *http.Cookie {
+func MakeCookieFromOptions(req *http.Request, name string, value string, opts *options.Cookie, expiration time.Duration, sameSite string) *http.Cookie {
 	domain := GetCookieDomain(req, opts.Domains)
 	// If nothing matches, create the cookie with the shortest domain
 	if domain == "" && len(opts.Domains) > 0 {
@@ -32,7 +32,7 @@ func MakeCookieFromOptions(req *http.Request, name string, value string, opts *o
 		Domain:   domain,
 		HttpOnly: opts.HTTPOnly,
 		Secure:   opts.Secure,
-		SameSite: ParseSameSite(opts.SameSite),
+		SameSite: ParseSameSite(sameSite),
 	}
 
 	if expiration > time.Duration(0) {