+++ pkg/cookies/cookies_suite_test.go
@@ -17,6 +17,10 @@ const (
 	cookieDomain = "o2p.cookies.test"
 	cookiePath   = "/cookie-tests"
 
+	sameSiteLax    = "lax"
+	sameSiteStrict = "strict"
+	sameSiteNone   = "none"
+
 	nowEpoch = 1609366421
 )