+++ pkg/cookies/csrf.go
@@ -134,6 +134,15 @@ func (c *csrf) SetSessionNonce(s *sessions.SessionState) {
 	s.Nonce = c.OIDCNonce
 }
 
+// getCSRFSameSite get the CSRF same site
+func getCSRFSameSite(opts *options.Cookie) string {
+	sameSite := opts.CSRFSameSite
+	if sameSite == "" {
+		sameSite = opts.SameSite
+	}
+	return sameSite
+}
+
 // SetCookie encodes the CSRF to a signed cookie and sets it on the ResponseWriter
 func (c *csrf) SetCookie(rw http.ResponseWriter, req *http.Request) (*http.Cookie, error) {
 	encoded, err := c.encodeCookie()
@@ -147,6 +156,7 @@ func (c *csrf) SetCookie(rw http.ResponseWriter, req *http.Request) (*http.Cooki
 		encoded,
 		c.cookieOpts,
 		c.cookieOpts.CSRFExpire,
+		getCSRFSameSite(c.cookieOpts),
 	)
 	http.SetCookie(rw, cookie)
 
@@ -203,6 +213,7 @@ func (c *csrf) ClearCookie(rw http.ResponseWriter, req *http.Request) {
 		"",
 		c.cookieOpts,
 		time.Hour*-1,
+		getCSRFSameSite(c.cookieOpts),
 	))
 }